pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lllm-core"
version = "0.0.1.3"
description = "Low-Level Language Model framework for building LLM agentic systems."
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Junyan Cheng" }]
requires-python = ">=3.10"
dependencies = [
    "openai>=1.40.0",
    "pydantic>=2.7.0",
    "tiktoken>=0.7.0",
    "numpy>=1.26.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "filelock>=3.13.0",
    "nbformat>=5.10.0",
    "jupyter-client>=8.6.0",
    "tomli>=2.0.1; python_version < \"3.11\"",
]

[project.scripts]
lllm = "lllm.cli:main"

[project.urls]
Homepage = "https://github.com/ChengJunyan1/lllm"
Source = "https://github.com/ChengJunyan1/lllm"
Documentation = "https://lllm.one"


[tool.setuptools.packages.find]
where = ["."]
include = ["lllm*"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools.package-data]
"lllm.core" = ["models.toml"]
"lllm.proxies.builtin" = ["*.json", "*.md"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "live: marks tests as live integration tests requiring real API keys (deselect with '-m not live')",
    "openai: marks tests that require OPENAI_API_KEY",
    "anthropic: marks tests that require ANTHROPIC_API_KEY",
]

